Added code1 to load method
Previously tests were failing with the error:
```KeyError: "['code1'] not in index"```
This was being caused when the `code1` column was not present in the data being input, as would be the case for new data downloaded from survey monkey. This commit adds a check for this in the laod method. If the variable is not present, an empty string variable will be added, whilst `code1` is added to the `raw.mapping` list.
`