Added additional tests
Various fixes including:
* Solve problem of code1 not being present in raw data:

Previously tests were failing with the error:
```KeyError: "['code1'] not in index"```
This was being caused when the `code1` column was not present in the data being input, as would be the case for new data downloaded from survey monkey. This commit adds a check for this in the laod method. If the variable is not present, an empty string variable will be added, whilst `code1` is added to the `raw.mapping` list.
`

* Added additional verbosity to method calls [temporary]
* Added new tests in  test_trainer.py, these check that:
    * code1 is present in survey.cleaned dataframe.
    * Conversion to OVA works as expected (for one and two classes).
* Updated setup.py